# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a RESTful ql-kho-lua.",
        "title": "Qua Ly Kho Lua",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/v1/api",
    "paths": {
...
        "/storehouses": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "get storehouses",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storehouses"
                ],
                "summary": "get storehouses",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Query",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Skip",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "minimum": 5,
                        "type": "integer",
                        "default": 5,
                        "description": "Limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Storehouses data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.responseWithPagination"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/handlers.userResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Create a new storehouse and get created storehouse data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storehouses"
                ],
                "summary": "Create a new storehouse and get created user data",
                "parameters": [
                    {
                        "description": "Create storehouse body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.createStorehouseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created storehouse data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.storehouseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicting data error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
        "/storehouses/{id}": {
            "get": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Get storehouse data by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storehouses"
                ],
                "summary": "Get storehouse data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Storehouse id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Created storehouse data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.storehouseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Delete a storehouse",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storehouses"
                ],
                "summary": "Delete a storehouse",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Storehouse id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated storehouse data",
                        "schema": {
                            "$ref": "#/definitions/handlers.response"
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "JWTAuth": []
                    }
                ],
                "description": "Update a storehouse and get created storehouse data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "storehouses"
                ],
                "summary": "Update a storehouse and get created user data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Storehouse id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update storehouse body",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.updateStorehouseRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated storehouse data",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/handlers.storehouseResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Validation error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "404": {
                        "description": "Data not found error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "409": {
                        "description": "Conflicting data error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/handlers.errorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "domain.Role": {
            "type": "string",
            "enum": [
                "root",
                "member"
            ],
            "x-enum-varnames": [
                "Root",
                "Member"
            ]
        },
...
        "handlers.createStorehouseRequest": {
            "type": "object",
            "required": [
                "capacity",
                "image",
                "location",
                "name"
            ],
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1200
                },
                "image": {
                    "type": "string",
                    "example": "2455.png"
                },
                "location": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        50.12,
                        68.36
                    ]
                },
                "name": {
                    "type": "string",
                    "minLength": 3,
                    "example": "store 01"
                }
            }
        },
...
        "handlers.errorResponse": {
            "type": "object",
            "properties": {
                "messages": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "data not found"
                    ]
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
...
        "handlers.pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer",
                    "example": 2
                },
                "limit_records": {
                    "type": "integer",
                    "example": 5
                },
                "next_page": {
                    "type": "integer",
                    "example": 2
                },
                "prev_page": {
                    "type": "integer",
                    "example": 1
                },
                "total_pages": {
                    "type": "integer",
                    "example": 10
                },
                "total_records": {
                    "type": "integer",
                    "example": 5
                }
            }
        },
        "handlers.response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "handlers.responseWithPagination": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string",
                    "example": "Success"
                },
                "pagination": {
                    "$ref": "#/definitions/handlers.pagination"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "handlers.storehouseResponse": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer",
                    "example": 1200
                },
                "id": {
                    "type": "integer"
                },
                "image": {
                    "type": "string",
                    "example": "2455.png"
                },
                "location": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        50.12,
                        68.36
                    ]
                },
                "name": {
                    "type": "string",
                    "example": "store 01"
                }
            }
        },
...
        "handlers.updateStorehouseRequest": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 1200
                },
                "image": {
                    "type": "string",
                    "example": "2455.png"
                },
                "location": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    },
                    "example": [
                        51.12,
                        68.36
                    ]
                },
                "name": {
                    "type": "string",
                    "minLength": 3,
                    "example": "store 01"
                }
            }
        },
...
        "handlers.userResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "example@exm.com"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "vertin"
                },
                "phone": {
                    "type": "string",
                    "example": "+84123456789"
                },
                "role": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/domain.Role"
                        }
                    ],
                    "example": "member"
                }
            }
        }
    },
    "securityDefinitions": {
        "JWTAuth": {
            "description": "Type \"JWT\" followed by a space and the access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}